Do not wait for a keypress when run as a service.
It may work well if the server waits for a key to press when run on a
desktop machine, but it will not work if it is run on a server.  On
servers, processes are often run in a non-interactive "headless" state
without a TTY attached.  That is why the application crashes there when
it tries to read from the standard input on launch.

To fix this, introduce a dedicated flag, `--daemon` to make this behavior
only optional.